Add Current singleton and refactor everything to work with it
