+++ app/controllers/concerns/set_host_url.rb
@@ -1,0 +1,13 @@
+module SetHostUrl
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :set_host_url
+  end
+
+  private
+
+  def set_host_url
+    Current.host_url = request.protocol + request.host
+  end
+end
