+++ app/controllers/users_controller.rb
@@ -34,6 +34,6 @@ class UsersController < ApplicationController
 
   def subscribe_to_repo(link)
     client = Octokit::Client.new access_token: session[:github_access_token]
-    client.subscribe("#{link}/events/push", @host_url + payload_path)
+    client.subscribe("#{link}/events/push", Current.host_url + payload_path)
   end
 end
