+++ app/views/task/_default.html.erb
@@ -5,18 +5,19 @@
   <%= task.text %>
 </p>
 
-<%= turbo_stream_from last_result_dom_id(task, @current_user) %>
 
-<% if @current_user.nil? %>
+<% if Current.user.nil? %>
   <div class="px-4 my-3 mb-5 py-5 bg-white shadow rounded-lg overflow-hidden sm:p-6">
     <dd class="mt-1 text-2xl font-semibold text-red-600">
       <%= "Для отображения результатов необходимо войти в аккаунт"%>
     </dd>
   </div>
 <% else %>
-  <div id="<%= last_result_dom_id(task, @current_user) %>">
-    <% if task.has_results_for_user?(@current_user) %>
-      <%= render task.last_result_for_user(@current_user) %>
+  <%= turbo_stream_from last_result_dom_id(task, Current.user) %>
+
+  <div id="<%= last_result_dom_id(task, Current.user) %>">
+    <% if task.has_results_for_user?(Current.user) %>
+      <%= render task.last_result_for_user(Current.user) %>
     <% else %>
       <div class="px-4 my-3 mb-5 py-5 bg-white shadow rounded-lg overflow-hidden sm:p-6">
         <dd class="mt-1 text-2xl font-semibold text-red-600">
